Added STM32I2c blanc driver controller

+++ interface/Log.h
@@ -3,7 +3,7 @@
 #ifdef AVR
 #include <stdint.h>
 #else
-#include <ctdint>
+#include <cstdint>
 #endif
 
 namespace driver
@@ -20,8 +20,8 @@ class  ILog
     };
 
     virtual ~ ILog() = default;
-    virtual void print(uint32_t channel, const char* message, ...) = 0;
-    virtual void value(uint32_t channel, int32_t value) = 0;
+    virtual void print(uint8_t channel, const char* message, ...) = 0;
+    virtual void value(uint8_t channel, int32_t value) = 0;
     virtual bool scan(char* string) = 0;
     virtual bool scan(int& number) = 0;
     virtual bool isInit() = 0;
